feat(deepseek-v41): add AutoModel training support
Integrate router replay, frozen Engram, CPU optimizer offload and refit support. Pin the corresponding AutoModel implementation and include the streaming optimizer probe.

+++ nemo_rl/models/generation/vllm/utils.py
@@ -140,7 +140,9 @@ def _as_routed_experts_tensor(
     global G_ROUTED_EXPERTS_RANGE_CHECKED
     tensor = torch.as_tensor(value, device=device)
     if not G_ROUTED_EXPERTS_RANGE_CHECKED and tensor.numel() > 0:
-        max_id = int(tensor.max())
+        # vLLM returns uint16 IDs for large expert pools; PyTorch does not
+        # implement max for uint16. Widen before checking, never narrow first.
+        max_id = int(tensor.to(torch.int64).max())
         limit = torch.iinfo(dtype).max
         if max_id > limit:
             raise ValueError(
